Cleaner way to load model and view from controller.
Using jQuery Deferred object to load dependent objects. This is much cleaner than figuring out for which objects should be loaded for called router function and then using if-else. Now loading of order and its line items are chained. So unless order is loaded, loading of lineitems is queued. If error occurs while loading order, line items will not be loaded.